feat(bcs-user-manager): support structured database config with cw-godb-sdk
- Add DatabaseConfig struct for mysql/postgres connection when DSN is empty
- Migrate from gorm v1 (github.com/jinzhu/gorm) to v2 (gorm.io/gorm)
- Integrate cw-godb-sdk for multi-database support (mysql, postgres)
- Maintain backward compatibility: DSN takes priority over DatabaseConfig
- Update connection pool parameters via DatabaseConfig (max_open_conns, max_idle_conns, conn_max_lifetime_second)
- Fix GORM v2 API changes: RemoveIndex->Migrator().DropIndex, Count->int64, DB() returns (sql.DB, error)

Configuration example:
{
  "database_config": {
    "db_type": "mysql",
    "db_host": "localhost",
    "db_port": 3306,
    "db_user": "root",
    "db_password": "encrypted_password",
    "db_name": "bcs",
    "max_open_conns": 20,
    "max_idle_conns": 20,
    "conn_max_lifetime_second": 60
  }
}

+++ bcs-services/bcs-user-manager/app/user-manager/storages/sqlstore/activity.go
@@ -44,14 +44,14 @@ func SearchActivities(projectCode, resourceType, activityType string, status mod
 	if endTime.Unix() != 0 {
 		query = query.Where("created_at <= ?", endTime)
 	}
-	count := 0
+	count := int64(0)
 	if err := query.Count(&count).Error; err != nil {
 		return nil, 0, err
 	}
 	if err := query.Offset(offset).Limit(limit).Order("created_at desc").Find(&activities).Error; err != nil {
 		return nil, 0, err
 	}
-	return activities, count, nil
+	return activities, int(count), nil
 }
 
 // CreateActivity create activity
